pyproject.toml
[project]
name = "tp-fastapi-graphql"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "uvicorn[standard]>=0.35.0,<0.36.0",
    "gunicorn>=23.0.0,<24.0.0",
    "pydantic>=2.11.7,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "httpx>=0.28.1,<0.29.0",
    "langchain>=0.3.27,<0.4.0",
    "langchain-openai>=0.3.35,<0.4.0",
    "strawberry-graphql>=0.281.0,<0.282.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1,<9.0.0",
    "pytest-cov>=6.2.1,<7.0.0",
    "pytest-asyncio>=0.25.0,<2.0.0",
    "pytest-mock>=3.14.1,<4.0.0",
    "aiosqlite>=0.21.0,<0.22.0",
    "coverage[toml]",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = "tests"
addopts = "-v -s"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
